refactor(docs): date-picker dx
