+++ apps/docs/content/components/date-picker/disabled.raw.jsx
@@ -1,0 +1,9 @@
+import {DatePicker} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <div className="flex w-full flex-wrap md:flex-nowrap gap-4">
+      <DatePicker isDisabled className="max-w-[284px]" label="Birth date" />
+    </div>
+  );
+}
