+++ apps/docs/content/components/date-picker/disabled.ts
@@ -1,16 +1,4 @@
-const App = `import {DatePicker} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex w-full flex-wrap md:flex-nowrap gap-4">
-        <DatePicker 
-          label="Birth date"
-          className="max-w-[284px]"
-          isDisabled
-        />
-    </div>
-  );
-}`;
+import App from "./disabled.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
