+++ apps/docs/content/components/date-picker/error-message-function.raw.jsx
@@ -1,0 +1,18 @@
+import {DatePicker} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <div className="flex w-full flex-wrap md:flex-nowrap gap-4">
+      <DatePicker
+        isInvalid
+        className="max-w-[284px]"
+        errorMessage={(value) => {
+          if (value.isInvalid) {
+            return "Please enter a valid date.";
+          }
+        }}
+        label="Birth date"
+      />
+    </div>
+  );
+}
