+++ apps/docs/content/components/date-picker/error-message-function.ts
@@ -1,21 +1,4 @@
-const App = `import {DatePicker} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex w-full flex-wrap md:flex-nowrap gap-4">
-      <DatePicker 
-        label="Birth date"
-        className="max-w-[284px]"
-        isInvalid
-        errorMessage={(value) => {
-          if (value.isInvalid) {
-            return "Please enter a valid date.";
-          }
-        }}
-      />
-    </div>
-  );
-}`;
+import App from "./error-message-funciton.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
