+++ apps/docs/content/components/date-picker/min-and-max-date.ts
@@ -1,28 +1,4 @@
-const App = `import {DatePicker} from "@nextui-org/react";
-import {getLocalTimeZone, today} from "@internationalized/date";
-
-export default function App() {
-  return (
-    <div className="w-full max-w-xl flex flex-row gap-4">
-      <div className="w-full flex flex-col gap-1">
-        <h3>Min date</h3>
-        <DatePicker
-          label="Date and time"
-          minValue={today(getLocalTimeZone())}
-          defaultValue={today(getLocalTimeZone()).subtract({ days: 1 })}
-        />
-      </div>
-      <div className="w-full flex flex-col gap-1">
-        <h3>Max date</h3>
-        <DatePicker
-          label="Date and time"
-          maxValue={today(getLocalTimeZone())}
-          defaultValue={today(getLocalTimeZone()).add({ days: 1 })}
-        />
-      </div>
-    </div>
-  );
-}`;
+import App from "./min-and-max-date.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
