+++ apps/docs/content/components/date-picker/page-behavior.ts
@@ -1,16 +1,4 @@
-const App = `import {DatePicker} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex w-full flex-wrap md:flex-nowrap gap-4">
-      <DatePicker 
-        label="Birth date"
-        visibleMonths={2}
-        pageBehavior="single"
-      />
-    </div> 
-  );
-}`;
+import App from "./page-behaviour.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
