+++ apps/docs/content/components/date-picker/selector-button-placement.ts
@@ -1,19 +1,4 @@
-const App = `import {DatePicker} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex w-full flex-wrap md:flex-nowrap gap-4">
-      <DatePicker 
-        label="Placement start"
-        selectorButtonPlacement="start"
-      />
-      <DatePicker 
-        label="Placement end (default)"
-        selectorButtonPlacement="end"
-      />
-    </div> 
-  );
-}`;
+import App from "./selector-button-placement.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
