+++ apps/docs/content/components/date-picker/unavailable-dates.raw.jsx
@@ -1,0 +1,30 @@
+import {DatePicker} from "@nextui-org/react";
+import {today, isWeekend, getLocalTimeZone} from "@internationalized/date";
+import {useLocale} from "@react-aria/i18n";
+
+export default function App() {
+  let now = today(getLocalTimeZone());
+
+  let disabledRanges = [
+    [now, now.add({days: 5})],
+    [now.add({days: 14}), now.add({days: 16})],
+    [now.add({days: 23}), now.add({days: 24})],
+  ];
+
+  let {locale} = useLocale();
+
+  let isDateUnavailable = (date) =>
+    isWeekend(date, locale) ||
+    disabledRanges.some(
+      (interval) => date.compare(interval[0]) >= 0 && date.compare(interval[1]) <= 0,
+    );
+
+  return (
+    <DatePicker
+      aria-label="Appointment date"
+      isDateUnavailable={isDateUnavailable}
+      label="Appointment date"
+      minValue={today(getLocalTimeZone())}
+    />
+  );
+}
