+++ apps/docs/content/components/date-picker/usage.raw.jsx
@@ -1,0 +1,5 @@
+import {DatePicker} from "@nextui-org/react";
+
+export default function App() {
+  return <DatePicker className="max-w-[284px]" label="Birth date" />;
+}
