+++ apps/docs/content/components/date-picker/usage.ts
@@ -1,10 +1,4 @@
-const App = `import {DatePicker} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <DatePicker label="Birth date" className="max-w-[284px]" />
-  );
-}`;
+import App from "./usage.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
