+++ apps/docs/content/components/date-picker/visible-month.raw.jsx
@@ -1,0 +1,9 @@
+import {DatePicker} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <div className="flex w-full flex-wrap md:flex-nowrap gap-4">
+      <DatePicker label="Birth date" visibleMonths={2} />
+    </div>
+  );
+}
