+++ apps/docs/content/components/date-picker/with-month-and-year-pickers.ts
@@ -1,16 +1,4 @@
-const App = `import {DatePicker} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="w-full max-w-xl flex flex-row gap-4">
-      <DatePicker
-        label="Birth Date"
-        variant="bordered"
-        showMonthAndYearPickers
-      />
-    </div>
-  );
-}`;
+import App from "./with-month-and-year-pickers.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
