+++ apps/docs/content/components/date-picker/with-time-field.ts
@@ -1,19 +1,4 @@
-const App = `import {DatePicker} from "@nextui-org/react";
-import {now, getLocalTimeZone} from "@internationalized/date";
-
-export default function App() {
-  return (
-    <div className="w-full max-w-xl flex flex-row gap-4">
-      <DatePicker
-        label="Event Date"
-        variant="bordered"
-        hideTimeZone
-        showMonthAndYearPickers
-        defaultValue={now(getLocalTimeZone())}
-      />
-    </div>
-  );
-}`;
+import App from "./with-time-field.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
